reset migrations and added dashboard models to its app
